Error in collision checks
Found the error

It seems like it was essentially - Sometimes the RRTs final point is in collision, and if this is the case (and it gets used for the starting point of the next call) then the planner crashes as it can't plan from a colliding spot.

I've added a warning message to mitigate this - possibly it should fail in this case rather than repairing it by replacing the path with 0,0 but that's a problem for later.

Fix error in implementation of resetjointstates

Applying start/end state checks to planners